backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Database
pymongo>=4.6.0
mongoengine>=0.27.0
motor>=3.3.0

# Async support
channels==4.0.0
channels-redis==4.2.0
asgiref==3.7.2

# AI/ML
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0
opencv-python==4.9.0.80
Pillow>=10.0.0
numpy<2.0,>=1.24.0

# Authentication & Security
PyJWT==2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0

# API integrations
requests>=2.31.0
httpx>=0.26.0

# File handling
python-multipart>=0.0.6

# Cache and Queue
redis>=5.0.0
django-redis>=5.4.0
celery>=5.3.0

# Monitoring
sentry-sdk[django]>=1.40.0

# Development
python-decouple>=3.8
pytest>=7.4.0
pytest-django>=4.8.0
pytest-asyncio>=0.23.0
coverage>=7.4.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.4
